+++ server/src/test/java/org/realityforge/gwt/appcache/server/PermutationsIOTest.java
@@ -29,7 +29,7 @@ public class PermutationsIOTest
       "   </permutation>\n" +
       "</permutations>\n";
 
-    final List<PermutationDescriptor> descriptors =
+    final List<SelectionDescriptor> descriptors =
       PermutationsIO.deserialize( new ByteArrayInputStream( permutationContent.getBytes( "US-ASCII" ) ) );
     assertEquals( descriptors.size(), 3 );
     final List<BindingProperty> binding1 = ensureBinding( descriptors, "C7D408F8EFA266A7F9A31209F8AA7446" );
@@ -65,24 +65,24 @@ public class PermutationsIOTest
     final String value22 = "smallish";
     final String value31 = "ie8,ie9,ie10,gecko1_8";
 
-    final ArrayList<PermutationDescriptor> input = new ArrayList<PermutationDescriptor>();
+    final ArrayList<SelectionDescriptor> input = new ArrayList<SelectionDescriptor>();
     final ArrayList<BindingProperty> input1 = new ArrayList<BindingProperty>();
     input1.add( new BindingProperty( key1, value11 ) );
     input1.add( new BindingProperty( key2, value12 ) );
     input1.add( new BindingProperty( key3, value13 ) );
-    input.add( new PermutationDescriptor( permutation1, input1 ) );
+    input.add( new SelectionDescriptor( permutation1, input1 ) );
 
     final ArrayList<BindingProperty> input2 = new ArrayList<BindingProperty>();
     input2.add( new BindingProperty( key1, value21 ) );
     input2.add( new BindingProperty( key2, value22 ) );
-    input.add( new PermutationDescriptor( permutation2, input2 ) );
+    input.add( new SelectionDescriptor( permutation2, input2 ) );
 
     final ArrayList<BindingProperty> input3 = new ArrayList<BindingProperty>();
     input3.add( new BindingProperty( key1, value31 ) );
-    input.add( new PermutationDescriptor( permutation3, input3 ) );
+    input.add( new SelectionDescriptor( permutation3, input3 ) );
 
     final String output = PermutationsIO.serialize( input );
-    final List<PermutationDescriptor> descriptors =
+    final List<SelectionDescriptor> descriptors =
       PermutationsIO.deserialize( new ByteArrayInputStream( output.getBytes( "US-ASCII" ) ) );
 
     assertEquals( descriptors.size(), 3 );
@@ -120,10 +120,10 @@ public class PermutationsIOTest
     return null;
   }
 
-  private List<BindingProperty> ensureBinding( final List<PermutationDescriptor> descriptors,
+  private List<BindingProperty> ensureBinding( final List<SelectionDescriptor> descriptors,
                                                final String permutationName )
   {
-    for ( final PermutationDescriptor descriptor : descriptors )
+    for ( final SelectionDescriptor descriptor : descriptors )
     {
       if ( descriptor.getPermutationName().equals( permutationName ) )
       {